Split home components
