+++ components/context.js
@@ -1,0 +1,34 @@
+import React, { useState } from 'react';
+
+export const AppContext = React.createContext();
+
+export const useApp = () => {
+  const context = React.useContext(AppContext);
+  if (!context) {
+    throw new Error('useApp must be used within AppProvider');
+  }
+  return context;
+};
+
+export const AppProvider = ({ children }) => {
+  const [currentPage, setCurrentPage] = useState('home');
+  const [user, setUser] = useState(null);
+  const [isAuthenticated, setIsAuthenticated] = useState(false);
+  const [searchResults, setSearchResults] = useState([]);
+  const [isLoading, setIsLoading] = useState(false);
+
+  const value = {
+    currentPage,
+    setCurrentPage,
+    user,
+    setUser,
+    isAuthenticated,
+    setIsAuthenticated,
+    searchResults,
+    setSearchResults,
+    isLoading,
+    setIsLoading
+  };
+
+  return <AppContext.Provider value={value}>{children}</AppContext.Provider>;
+};
